Add Jonathan Huckabee Games Portfolio
- Created portfolio page showcasing Jonathan Huckabee's video game development work
- Added developer profile with bio, skills, and links to Cyan Web Solutions and The Streets Network
- Implemented game showcase featuring Squirrelmageddon (published on Steam)
- Set up Angular routing to support both Bargain Dispatch (deals) and Games Portfolio sections
- Refactored app structure:
  - Main app component now includes navigation header and footer
  - Moved existing deals functionality to dedicated component
  - Created reusable game card component for portfolio display
- Added TypeScript models for Developer and Game data
- Created JSON data files for easy content management
- Implemented PortfolioService to load game and developer data
- Maintained vintage newspaper aesthetic across the application
- Build successfully tested and verified

Jonathan Huckabee is the Director & Founder of Cyan Web Solutions with 30+ years of IT experience,
and founder of The Streets Network which published Squirrelmageddon on Steam.
