+++ src/app/app.config.ts
@@ -1,6 +1,11 @@
 import { ApplicationConfig } from '@angular/core';
 import { provideHttpClient } from '@angular/common/http';
+import { provideRouter } from '@angular/router';
+import { routes } from './app.routes';
 
 export const appConfig: ApplicationConfig = {
-  providers: [provideHttpClient()]
+  providers: [
+    provideHttpClient(),
+    provideRouter(routes)
+  ]
 };
