+++ src/app/components/game-card/game-card.component.scss
@@ -1,0 +1,111 @@
+.game-card {
+  background: #fff;
+  border-radius: 8px;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+  overflow: hidden;
+  transition: transform 0.2s ease, box-shadow 0.2s ease;
+
+  &:hover {
+    transform: translateY(-4px);
+    box-shadow: 0 4px 16px rgba(0, 0, 0, 0.15);
+  }
+}
+
+.game-image {
+  width: 100%;
+  height: 200px;
+  overflow: hidden;
+  background: #f0f0f0;
+
+  img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+  }
+}
+
+.game-content {
+  padding: 1.5rem;
+}
+
+.game-title {
+  margin: 0 0 0.5rem 0;
+  font-size: 1.5rem;
+  color: #333;
+}
+
+.game-meta {
+  color: #666;
+  font-size: 0.9rem;
+  margin-bottom: 1rem;
+}
+
+.game-description {
+  color: #555;
+  line-height: 1.6;
+  margin-bottom: 1rem;
+}
+
+.game-genres {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.5rem;
+  margin-bottom: 1rem;
+}
+
+.genre-tag {
+  background: #e0e0e0;
+  padding: 0.25rem 0.75rem;
+  border-radius: 16px;
+  font-size: 0.85rem;
+  color: #555;
+}
+
+.game-platforms,
+.game-release {
+  font-size: 0.9rem;
+  color: #666;
+  margin-bottom: 0.5rem;
+
+  strong {
+    color: #333;
+  }
+}
+
+.game-features {
+  margin: 1rem 0;
+  font-size: 0.9rem;
+
+  strong {
+    color: #333;
+  }
+
+  ul {
+    margin: 0.5rem 0 0 0;
+    padding-left: 1.5rem;
+    color: #666;
+
+    li {
+      margin-bottom: 0.25rem;
+    }
+  }
+}
+
+.game-actions {
+  margin-top: 1.5rem;
+
+  .steam-btn {
+    background: #171a21;
+    color: #fff;
+    border: none;
+    padding: 0.75rem 1.5rem;
+    border-radius: 4px;
+    font-size: 1rem;
+    cursor: pointer;
+    transition: background 0.2s ease;
+
+    &:hover {
+      background: #2a475e;
+    }
+  }
+}
