+++ src/app/layout/pages/portfolio/portfolio.component.html
@@ -1,0 +1,62 @@
+<div class="portfolio-page">
+  <div class="loading-spinner" *ngIf="loading">
+    <p>Loading portfolio...</p>
+  </div>
+
+  <div class="portfolio-content" *ngIf="!loading">
+    <!-- Developer Profile Section -->
+    <section class="developer-profile" *ngIf="developer">
+      <div class="profile-header">
+        <div class="profile-image" *ngIf="developer.imageUrl">
+          <img [src]="developer.imageUrl" [alt]="developer.name" />
+        </div>
+        <div class="profile-info">
+          <h1>{{ developer.name }}</h1>
+          <h2>{{ developer.title }} - {{ developer.company }}</h2>
+          <p class="location">{{ developer.location }}</p>
+          <p class="experience">{{ developer.experience }}</p>
+        </div>
+      </div>
+
+      <div class="profile-bio">
+        <p>{{ developer.bio }}</p>
+      </div>
+
+      <div class="profile-skills" *ngIf="developer.skills && developer.skills.length > 0">
+        <h3>Skills & Expertise</h3>
+        <div class="skills-list">
+          <span class="skill-tag" *ngFor="let skill of developer.skills">{{ skill }}</span>
+        </div>
+      </div>
+
+      <div class="profile-links" *ngIf="developer.websites && developer.websites.length > 0">
+        <h3>Websites</h3>
+        <ul>
+          <li *ngFor="let website of developer.websites">
+            <a [href]="website.url" target="_blank">{{ website.label }}</a>
+          </li>
+        </ul>
+      </div>
+
+      <div class="profile-social" *ngIf="developer.socialLinks && developer.socialLinks.length > 0">
+        <h3>Connect</h3>
+        <div class="social-links">
+          <a *ngFor="let link of developer.socialLinks" [href]="link.url" target="_blank" class="social-link">
+            {{ link.platform }}
+          </a>
+        </div>
+      </div>
+    </section>
+
+    <!-- Games Section -->
+    <section class="games-section">
+      <h2>Published Video Games</h2>
+      <div class="games-grid" *ngIf="games.length > 0">
+        <app-game-card *ngFor="let game of games" [game]="game"></app-game-card>
+      </div>
+      <div class="no-games" *ngIf="games.length === 0">
+        <p>No games available at this time.</p>
+      </div>
+    </section>
+  </div>
+</div>
